Move push_admin out of src
