+++ push_admin/messaging.py
@@ -14,8 +14,9 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from src.push_admin import _messages, _app
-from src import push_admin
+from push_admin import _messages
+from push_admin import _app
+import push_admin
 
 """HUAWEI Cloud Messaging module."""
