+++ setup.py
@@ -39,7 +39,7 @@ setup(
     license='Apache License 2.0',
     keywords='huawei cloud development',
     install_requires=install_requires,
-    packages=['src/push_admin'],
+    packages=['push_admin'],
     python_requires='>=2.7, !=3.0.*, !=3.1.*, !=3.2.*, !=3.3.*',
     classifiers=[
         'Development Status :: 5 - Production/Stable',
